feat(social): SocialCapitalView MV reducer over social/* Datoms (ADR-2606082100 §5)
Wires the social-capital ledger into kotoba's incremental MaterializedView path —
the economic loop now runs over the Datom log. Geometric half-life decay is not
Datalog-expressible, so SocialCapitalView is a custom reducer sibling of
mv::MaterializedView: feed it each commit's Deltas via apply(), read a DID's
decayed capital via capital() or a root's SC_root via capital_sum() (the retainer
allocation numerator, spec §6).

Architecture: moved the decay primitives + SocialCapitalLedger from kotoba-server
to kotoba-kqe (the query/MV engine) so the reducer and the ledger share ONE decay
primitive; kotoba-server::social is now a thin re-export (path preserved, callers
unaffected). No circular dep (server → kqe).

- crates/kotoba-kqe/src/social.rs: primitives + ledger + SocialCapitalView + 20 tests
  (parse, mint accumulate+decay, disclosure+wellbecoming, burn clamp, ignore
  non-social/retracts, capital_sum SC_root, unknown-DID zero).
- kotoba-server/src/social.rs: `pub use kotoba_kqe::social::*;`.

kqe 20 social tests green; server 416 green; clippy-clean.

Follow-ups: MvRegistry generalization for custom reducers; mint job feeding Deltas
from observed anchor-chain events + CitationLedger + KaizenObserver; XRPC read.
